+++ pkg/analysis_server/lib/src/services/correction/dart/create_getter.dart
@@ -141,7 +141,7 @@ class CreateGetter extends CreateFieldOrGetter {
     var staticModifier = false;
     Element? targetElement;
     if (target is ExtensionOverride) {
-      targetElement = target.staticElement;
+      targetElement = target.element;
     } else if (target is Identifier &&
         target.staticElement is ExtensionElement) {
       targetElement = target.staticElement;