+++ pkg/analyzer/test/src/diagnostics/undefined_extension_method_test.dart
@@ -40,16 +40,14 @@ f() {
     assertResolvedNodeText(node, r'''
 MethodInvocation
   target: ExtensionOverride
-    extensionName: SimpleIdentifier
-      token: E
-      staticElement: self::@extension::E
-      staticType: null
+    name: E
     argumentList: ArgumentList
       leftParenthesis: (
       arguments
         SimpleStringLiteral
           literal: 'a'
       rightParenthesis: )
+    element: self::@extension::E
     extendedType: String
     staticType: null
   operator: .